requirements.txt
Django==3.2
django-filter==23.2
djangorestframework==3.14.0
drf-spectacular==0.26.2
Markdown==3.4.3
scikit-learn==1.3.0
scipy==1.11.1
psycopg2-binary==2.9.1
